+++ test/src/validation_test.dart
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import 'package:betto_common/collections.dart' show Range;
 import 'package:betto_schema/betto_schema.dart';
-import 'package:betto_schema/src/range.dart';
 
 import 'package:test/test.dart';